Add code notes for 6A lecture about streams
The code is functional, I make some changes for this works on Common Lisp.
The major problem is store the cdr of stream as non-evaluated expression to
execute that in future. But using some black magic of quotation on the
internal procedures for streams this works.

I need quote the code for (cons-stream (head s) `(filter ,#'oddp ,(tail s))
But using quasi-quote because I need the binding values for each variable.