Watch lecture 6B about others procedures of streams
I don't added the code of this because doesn't works well, as commented
inside of file. That lecture introduce some interesting concepts of evaluation
like the problem I described for don't evaluate the arguments until you need.
The name for this is normal-order evaluation. A great example of this nowdays
is the Haskell language.

Common Lisp, as well Scheme, don't use normal-order evaluation. That way
we can have trouble about avoiding evaluation of arguments to create streams
and promises.