Add notes for 8B lecture (SICP)
The implementation on the Logic Programming II of the query
language, rule-based using pattern matching makes assumptions about
earlier codes from old lectures. I just keep a general copy-paste
notes from the videos make my own reasoning.

Is curious case. The NOT from a query rule-based system is
different from the NOT LOGICAL. The closed system doesn't build
all the knowledge possible. The NOT from query is a filter, making
a equivalent procedure for the NOT operation on Set Theory, the
complements of a set.

Hmmm... consistency on a closed world. Maybe is this was the initial
insights with Godel on the Incompletude Theorem?

... Who cares.