Finished the book Land of Lisp!
A great book for introduction of Lisp. The first chapters is really great!
Some examples are easy to understand, but others are not. I think using
Dice of Doom on almost 4 chapters is not a very good idea. BTW, the style
of load/redefining functions is silly.

But keeping this bad things on other side, this would a great journey of
the world of Common Lisp. Great learning.

I'll now focus into the Discrete Mathematics with Haskell on the repository
www.github.com/ryukinix/discrete-mathematics.